format R128 GAIN tags as Q7.8 number
Save the value in R128_{TRACK,ALBUM}_GAIN tags not as a decimal float but as a
Q7.8 number as demanded by IETF RFC 7845, 5.3.1.

This implementation actually saves a Q8 number, i.e. does not check if the
integer part fits into 7 bits. While reading this is clearly the most sensible
approach. We could consider adding a check when writing and issuing either a
Warning or an Error.