Add support for splitting of ID3v2.3 multi-value tags for MP3ListDescStorageStyle
fixes rest of problems in #43

Shouldn't be enabled on tags like `artists` or other without confidence that values will not contain separator (eg. `artists: ['AC/DC']`).
Format of mbid is known and I'm sure we can enable it here.

When solving #21 this separator value in `el.split('/')` also should be taken into account.